# transport: Optimize allocations