This change uses a sync pool to cache `itemNodes` used by `controlbuf` and stores a method as a closure to avoid a heap allocation while creating frames for the `controlbuf`. This brings down `allocs/op` by ~18%.

```go
❯ go run benchmark/benchresult/main.go before after
streaming-networkMode_Local-bufConn_false-keepalive_false-benchTime_30s-trace_false-latency_0s-kbps_0-MTU_0-maxConcurrentCalls_60-reqS
ize_32384B-respSize_32384B-compressor_off-channelz_false-preloader_false-clientReadBufferSize_-1-clientWriteBufferSize_-1-serverReadBu
fferSize_-1-serverWriteBufferSize_-1-sleepBetweenRPCs_0s-connections_1-recvBufferPool_simple-sharedWriteBuffer_false
               Title       Before        After Percentage
            TotalOps      1089062      1092837     0.35%
             SendOps            0            0      NaN%
             RecvOps            0            0      NaN%
            Bytes/op    135783.55    135773.63    -0.01%
           Allocs/op        32.20        26.33   -18.63%
             ReqT/op 9404849015.47 9437448908.80     0.35%
            RespT/op 9404849015.47 9437448908.80     0.35%
            50th-Lat   1.646019ms   1.639128ms    -0.42%
            90th-Lat   2.280214ms   2.261793ms    -0.81%
            99th-Lat   2.692703ms   2.678714ms    -0.52%
             Avg-Lat   1.651539ms    1.64601ms    -0.33%
           GoVersion     go1.24.2     go1.24.2
         GrpcVersion   1.74.0-dev   1.74.0-dev
```

RELEASE NOTES:
* transport: reduce heap allocations by pooling objects and avoiding method-to-closure conversions.